댓글 삭제 및 수정 기능 추가

+++ src/main/java/com/climbing/climbing_web/dto/CommentDTO.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class CommentDTO {
-    private Long commentId;
+    private int commentId;
     private int postId;
     private String memberId;
     private String writerName;
